FIT-81-Domingo 30

+++ src/main/java/com/fitbum/entidades/Notificacion.java
@@ -21,7 +21,7 @@ public class Notificacion {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 
-    private Integer idNotificacion;
+    private Long idNotificacion;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     @JsonFormat(pattern="yyyy-MM-dd")
